@@ -39,47 +39,47 @@
     "karma-source-map-support": "1.4.0",
     "less": "4.1.2",
     "less-loader": "10.2.0",
-    "license-webpack-plugin": "4.0.0",
+    "license-webpack-plugin": "4.0.1",
     "loader-utils": "3.2.0",
-    "mini-css-extract-plugin": "2.4.5",
+    "mini-css-extract-plugin": "2.5.3",
     "minimatch": "3.0.4",
     "open": "8.4.0",
     "ora": "5.4.1",
     "parse5-html-rewriting-stream": "6.0.1",
-    "piscina": "3.1.0",
-    "postcss": "8.4.4",
+    "piscina": "3.2.0",
+    "postcss": "8.4.5",
     "postcss-import": "14.0.2",
     "postcss-loader": "6.2.1",
     "postcss-preset-env": "7.2.3",
     "regenerator-runtime": "0.13.9",
-    "resolve-url-loader": "4.0.0",
+    "resolve-url-loader": "5.0.0",
     "rxjs": "6.6.7",
-    "sass": "1.44.0",
+    "sass": "1.49.0",
     "sass-loader": "12.4.0",
     "semver": "7.3.5",
-    "source-map-loader": "3.0.0",
+    "source-map-loader": "3.0.1",
     "source-map-support": "0.5.21",
-    "stylus": "0.55.0",
+    "stylus": "0.56.0",
     "stylus-loader": "6.2.0",
     "terser": "5.10.0",
     "text-table": "0.2.0",
     "tree-kill": "1.2.2",
     "tslib": "2.3.1",
-    "webpack": "5.65.0",
-    "webpack-dev-middleware": "5.2.2",
-    "webpack-dev-server": "4.6.0",
+    "webpack": "5.67.0",
+    "webpack-dev-middleware": "5.3.0",
+    "webpack-dev-server": "4.7.3",
     "webpack-merge": "5.8.0",
-    "webpack-subresource-integrity": "5.0.0"
+    "webpack-subresource-integrity": "5.1.0"
   },
   "optionalDependencies": {
-    "esbuild": "0.14.11"
+    "esbuild": "0.14.14"
   },
   "peerDependencies": {
-    "@angular/compiler-cli": "^13.0.0 || ^13.1.0-next",
-    "@angular/localize": "^13.0.0 || ^13.1.0-next",
-    "@angular/service-worker": "^13.0.0 || ^13.1.0-next",
+    "@angular/compiler-cli": "^13.0.0",
+    "@angular/localize": "^13.0.0",
+    "@angular/service-worker": "^13.0.0",
     "karma": "^6.3.0",
-    "ng-packagr": "^13.0.0 || ^13.1.0-next",
+    "ng-packagr": "^13.0.0",
     "protractor": "^7.0.0",
     "tailwindcss": "^2.0.0 || ^3.0.0",
     "typescript": ">=4.4.3 <4.6"
