@@ -1,6 +1,6 @@
 {
   "name": "@jridgewell/resolve-uri",
-  "version": "1.0.0",
+  "version": "3.0.5",
   "description": "Resolve a URI relative to an optional base URI",
   "keywords": [
     "resolve",
@@ -8,59 +8,57 @@
     "url",
     "path"
   ],
+  "author": "Justin Ridgewell <justin@ridgewell.name>",
+  "license": "MIT",
+  "repository": "https://github.com/jridgewell/resolve-uri",
   "main": "dist/resolve-uri.umd.js",
   "module": "dist/resolve-uri.mjs",
   "typings": "dist/types/resolve-uri.d.ts",
+  "exports": {
+    ".": {
+      "browser": "./dist/resolve-uri.umd.js",
+      "require": "./dist/resolve-uri.umd.js",
+      "import": "./dist/resolve-uri.mjs"
+    },
+    "./package.json": "./package.json"
+  },
   "files": [
     "dist"
   ],
-  "author": "Justin Ridgewell <jridgewell@google.com>",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jridgewell/resolve-uri.git"
-  },
-  "bugs": {
-    "url": "https://github.com/jridgewell/resolve-uri/issues"
-  },
-  "license": "MIT",
   "engines": {
     "node": ">=6.0.0"
   },
   "scripts": {
-    "lint:ts": "npm run test:lint:ts -- --fix",
-    "lint:prettier": "npm run test:lint:prettier -- --write",
-    "lint": "run-s -n lint:*",
     "prebuild": "rm -rf dist",
-    "build:ts": "tsc --module commonjs",
-    "build:rollup": "rollup -c rollup.config.ts",
     "build": "run-s -n build:*",
-    "test": "jest --coverage",
-    "test:debug": "node --inspect-brk node_modules/.bin/jest --runInBand",
-    "test:watch": "jest --coverage --watch",
-    "test:lint:ts": "tslint  --project tsconfig.json -t codeFrame '{src,test}/**/*.ts'",
-    "test:lint:prettier": "prettier --check '{src,test}/**/*.ts'",
+    "build:rollup": "rollup -c rollup.config.js",
+    "build:ts": "tsc --project tsconfig.build.json",
+    "lint": "run-s -n lint:*",
+    "lint:prettier": "npm run test:lint:prettier -- --write",
+    "lint:ts": "npm run test:lint:ts -- --fix",
+    "pretest": "run-s build:rollup",
+    "test": "run-s -n test:lint test:only",
+    "test:debug": "mocha --inspect-brk",
     "test:lint": "run-s -n test:lint:*",
-    "test:prod": "run-s -n test:lint 'test --no-cache'",
-    "preversion": "run-s test:prod build",
-    "prepublishOnly": "npm run preversion"
+    "test:lint:prettier": "prettier --check '{src,test}/**/*.ts'",
+    "test:lint:ts": "eslint '{src,test}/**/*.ts'",
+    "test:only": "mocha",
+    "test:coverage": "c8 mocha",
+    "test:watch": "mocha --watch",
+    "prepublishOnly": "npm run preversion",
+    "preversion": "run-s test build"
   },
   "devDependencies": {
-    "@types/jest": "24.0.15",
-    "@types/node": "12.6.2",
-    "jest": "24.8.0",
-    "jest-config": "24.8.0",
+    "@rollup/plugin-typescript": "8.3.0",
+    "@typescript-eslint/eslint-plugin": "5.10.0",
+    "@typescript-eslint/parser": "5.10.0",
+    "c8": "7.11.0",
+    "eslint": "8.7.0",
+    "eslint-config-prettier": "8.3.0",
+    "mocha": "9.2.0",
     "npm-run-all": "4.1.5",
-    "prettier": "1.18.2",
-    "rollup": "1.16.7",
-    "rollup-plugin-commonjs": "10.0.1",
-    "rollup-plugin-node-resolve": "5.2.0",
-    "rollup-plugin-sourcemaps": "0.4.2",
-    "rollup-plugin-typescript": "1.0.1",
-    "ts-jest": "24.0.2",
-    "ts-node": "8.3.0",
-    "tslint": "5.18.0",
-    "tslint-config-prettier": "1.18.0",
-    "tslint-config-standard": "8.0.1",
-    "typescript": "3.5.3"
+    "prettier": "2.5.1",
+    "rollup": "2.66.0",
+    "typescript": "4.5.5"
   }
 }
