@@ -1,6 +1,6 @@
 {
   "name": "are-we-there-yet",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "description": "Keep track of the overall completion of many disparate processes",
   "main": "lib/index.js",
   "scripts": {
@@ -13,7 +13,9 @@
     "preversion": "npm test",
     "postversion": "npm publish",
     "prepublishOnly": "git push origin --follow-tags",
-    "snap": "tap"
+    "snap": "tap",
+    "postlint": "npm-template-check",
+    "template-copy": "npm-template-copy --force"
   },
   "repository": {
     "type": "git",
@@ -26,9 +28,9 @@
   },
   "homepage": "https://github.com/npm/are-we-there-yet",
   "devDependencies": {
-    "@npmcli/eslint-config": "^1.0.0",
-    "@npmcli/template-oss": "^1.0.2",
-    "eslint": "^7.32.0",
+    "@npmcli/eslint-config": "^2.0.0",
+    "@npmcli/template-oss": "^2.7.1",
+    "eslint": "^8.8.0",
     "eslint-plugin-node": "^11.1.0",
     "tap": "^15.0.9"
   },
@@ -41,7 +43,7 @@
     "lib"
   ],
   "engines": {
-    "node": ">=10"
+    "node": "^12.13.0 || ^14.15.0 || >=16"
   },
   "tap": {
     "branches": 68,
@@ -49,5 +51,7 @@
     "functions": 86,
     "lines": 92
   },
-  "templateVersion": "1.0.2"
+  "templateOSS": {
+    "version": "2.7.1"
+  }
 }
