@@ -1,6 +1,6 @@
 {
   "name": "mini-css-extract-plugin",
-  "version": "2.4.5",
+  "version": "2.5.3",
   "description": "extracts CSS into separate files",
   "license": "MIT",
   "repository": "webpack-contrib/mini-css-extract-plugin",
@@ -11,19 +11,23 @@
     "type": "opencollective",
     "url": "https://opencollective.com/webpack"
   },
-  "main": "dist/cjs.js",
+  "main": "dist/index.js",
+  "types": "types/index.d.ts",
   "engines": {
     "node": ">= 12.13.0"
   },
   "scripts": {
     "start": "npm run build -- -w",
     "prebuild": "npm run clean",
-    "build": "cross-env NODE_ENV=production babel src -d dist --ignore \"src/**/*.test.js\" --copy-files",
+    "build:types": "tsc --declaration --emitDeclarationOnly --outDir types && prettier \"types/**/*.ts\" --write",
+    "build:code": "cross-env NODE_ENV=production babel src -d dist --copy-files",
+    "build": "npm-run-all -p \"build:**\"",
     "postbuild": "es-check es5 dist/hmr/hotModuleReplacement.js",
     "clean": "del-cli dist",
     "commitlint": "commitlint --from=master",
     "lint:prettier": "prettier \"{**/*,*}.{js,json,md,yml,css,ts}\" --list-different",
     "lint:js": "eslint --cache .",
+    "lint:types": "tsc --pretty --noEmit",
     "lint": "npm-run-all -l -p \"lint:**\"",
     "prepare": "husky install && npm run build",
     "release": "standard-version",
@@ -37,7 +41,8 @@
     "test": "cross-env NODE_ENV=test npm run test:coverage"
   },
   "files": [
-    "dist"
+    "dist",
+    "types"
   ],
   "peerDependencies": {
     "webpack": "^5.0.0"
@@ -46,10 +51,10 @@
     "schema-utils": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.5",
-    "@babel/core": "^7.14.6",
+    "@babel/cli": "^7.16.7",
+    "@babel/core": "^7.16.7",
     "@babel/eslint-parser": "^7.16.0",
-    "@babel/preset-env": "^7.14.9",
+    "@babel/preset-env": "^7.16.7",
     "@commitlint/cli": "^15.0.0",
     "@commitlint/config-conventional": "^15.0.0",
     "@webpack-contrib/eslint-config-webpack": "^3.0.0",
@@ -60,23 +65,24 @@
     "del": "^6.0.0",
     "del-cli": "^4.0.0",
     "es-check": "^6.0.0",
-    "eslint": "^8.1.0",
+    "eslint": "^8.6.0",
     "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-import": "^2.25.4",
     "file-loader": "^6.2.0",
     "husky": "^7.0.0",
     "jest": "^27.0.6",
-    "jsdom": "^18.0.0",
-    "lint-staged": "^12.0.2",
-    "memfs": "^3.0.2",
+    "jsdom": "^19.0.0",
+    "lint-staged": "^12.1.5",
+    "memfs": "^3.4.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.2",
-    "sass": "^1.39.0",
+    "sass": "^1.45.2",
     "sass-loader": "^12.1.0",
     "standard-version": "^9.3.0",
+    "typescript": "^4.5.4",
     "webpack": "^5.58.1",
     "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^4.0.0"
+    "webpack-dev-server": "^4.7.2"
   },
   "keywords": [
     "webpack",
