@@ -1,6 +1,6 @@
 {
   "name": "piscina",
-  "version": "3.1.0",
+  "version": "3.2.0",
   "description": "A fast, efficient Node.js Worker Thread Pool implementation",
   "main": "./dist/src/index.js",
   "exports": {
@@ -42,7 +42,7 @@
     "standardx": "^7.0.0",
     "tap": "^15.0.6",
     "ts-node": "^9.1.1",
-    "typescript": "^4.2.4"
+    "typescript": "4.3.x"
   },
   "dependencies": {
     "eventemitter-asyncresource": "^1.0.0",
