@@ -1,6 +1,6 @@
 {
   "name": "resolve-url-loader",
-  "version": "4.0.0",
+  "version": "5.0.0",
   "description": "Webpack loader that resolves relative paths in url() statements based on the original source file",
   "main": "index.js",
   "repository": {
@@ -25,9 +25,9 @@
   "bugs": {
     "url": "https://github.com/bholloway/resolve-url-loader/issues"
   },
-  "homepage": "https://github.com/bholloway/resolve-url-loader/tree/v4-maintenance/packages/resolve-url-loader",
+  "homepage": "https://github.com/bholloway/resolve-url-loader/tree/v5/packages/resolve-url-loader",
   "engines": {
-    "node": ">=8.9"
+    "node": ">=12"
   },
   "files": [
     "index.js",
@@ -38,19 +38,7 @@
     "adjust-sourcemap-loader": "^4.0.0",
     "convert-source-map": "^1.7.0",
     "loader-utils": "^2.0.0",
-    "postcss": "^7.0.35",
+    "postcss": "^8.2.14",
     "source-map": "0.6.1"
-  },
-  "peerDependencies": {
-    "rework": "1.0.1",
-    "rework-visit": "1.0.0"
-  },
-  "peerDependenciesMeta": {
-    "rework": {
-      "optional": true
-    },
-    "rework-visit": {
-      "optional": true
-    }
   }
 }
