@@ -1,7 +1,7 @@
 {
   "name": "socks",
   "private": false,
-  "version": "2.6.1",
+  "version": "2.6.2",
   "description": "Fully featured SOCKS proxy client supporting SOCKSv4, SOCKSv4a, and SOCKSv5. Includes Bind and Associate functionality.",
   "main": "build/index.js",
   "typings": "typings/index.d.ts",
@@ -34,21 +34,19 @@
   "readmeFilename": "README.md",
   "devDependencies": {
     "@types/ip": "1.1.0",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^14.14.41",
-    "coveralls": "3.1.0",
-    "mocha": "^8.3.2",
+    "@types/mocha": "^9.1.0",
+    "@types/node": "^17.0.15",
+    "coveralls": "^3.1.1",
+    "mocha": "^9.2.0",
     "nyc": "15.1.0",
-    "prettier": "^2.2.1",
+    "prettier": "^2.5.1",
     "socks5-server": "^0.1.1",
-    "ts-node": "^9.1.1",
-    "tslint": "^6.1.3",
-    "tslint-config-airbnb": "^5.11.2",
-    "typescript": "^4.2.4"
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.5"
   },
   "dependencies": {
     "ip": "^1.1.5",
-    "smart-buffer": "^4.1.0"
+    "smart-buffer": "^4.2.0"
   },
   "scripts": {
     "prepublish": "npm install -g typescript && npm run build",
